Add files via upload
This is a webpage that produce or give sounds when some certain keys or the buttons made available are pressed on the page. the buttons / keys are specific to a particular sound, frankly its a beat producing webpage ( chuckles! ). Languages used are html for the outlay, some css for the styling to make it more appealing, could have add some background images but couldn't find any suitable ones to compliment each sounds and lastly some javascript for user interface ease, i.e the button and key linking on the website.